# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Sistema de Gestao para as Oficinas Gengis Khan API
  version: 0.1.0
  description: API machine-readable baseline aligned with backend implementation.
servers:
  - url: http://localhost:3000
...
paths:
...
  /api/v1/service-orders:
    get:
      tags: [service-orders]
      summary: List service orders
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Service orders list
    post:
      tags: [service-orders]
      summary: Create service order
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceOrderCreateRequest'
      responses:
        '200':
          description: Created service order
  /api/v1/service-orders/summary:
    get:
      tags: [service-orders]
      summary: Get service orders summary
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Summary by status
  /api/v1/service-orders/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      tags: [service-orders]
      summary: Get service order by id
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Service order detail
    patch:
      tags: [service-orders]
      summary: Update service order editable fields
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Updated service order
  /api/v1/service-orders/{id}/status:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    patch:
      tags: [service-orders]
      summary: Transition service order status
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [toStatus]
              properties:
                toStatus:
                  type: string
                note:
                  type: string
      responses:
        '200':
          description: Updated status
  /api/v1/service-orders/{id}/diagnosis:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    patch:
      tags: [service-orders]
      summary: Update diagnosis
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Updated service order diagnosis
  /api/v1/service-orders/{id}/budget:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    patch:
      tags: [service-orders]
      summary: Update budget
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Updated budget
  /api/v1/service-orders/{id}/budget/approval:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    post:
      tags: [service-orders]
      summary: Record budget approval or rejection
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Budget approval registered
  /api/v1/service-orders/{id}/history:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      tags: [service-orders]
      summary: Get service order history
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Service order history
...
  /api/v1/suppliers:
    get:
      tags: [suppliers-procurement]
      summary: List suppliers
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Suppliers list
    post:
      tags: [suppliers-procurement]
      summary: Create supplier
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Supplier created
  /api/v1/suppliers/{supplierId}:
    parameters:
      - name: supplierId
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [suppliers-procurement]
      summary: Get supplier
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Supplier detail
    patch:
      tags: [suppliers-procurement]
      summary: Update supplier
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Supplier updated
  /api/v1/purchase-orders:
    get:
      tags: [suppliers-procurement]
      summary: List purchase orders
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Purchase order list
    post:
      tags: [suppliers-procurement]
      summary: Create purchase order
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Purchase order created
  /api/v1/purchase-orders/generate-from-low-stock:
    post:
      tags: [suppliers-procurement]
      summary: Generate purchase orders from low stock
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Purchase orders generated
  /api/v1/purchase-orders/{purchaseOrderId}:
    parameters:
      - name: purchaseOrderId
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [suppliers-procurement]
      summary: Get purchase order
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Purchase order detail
    patch:
      tags: [suppliers-procurement]
      summary: Update purchase order
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Purchase order updated
  /api/v1/purchase-orders/{purchaseOrderId}/receive:
    parameters:
      - name: purchaseOrderId
        in: path
        required: true
        schema:
          type: string
    patch:
      tags: [suppliers-procurement]
      summary: Mark purchase order as received
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Purchase order received
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
...
  schemas:
...
    ServiceOrderCreateRequest:
      type: object
      required: [customerNif, scooterSerialNumber, reportedProblem]
      properties:
        customerNif:
          type: string
        scooterSerialNumber:
          type: string
        reportedProblem:
          type: string
        estimatedCompletionDate:
          type: string
          format: date